Qrcode for order info

+++ src/app/common/Image.php
@@ -8,26 +8,31 @@ use Picqer\Barcode\BarcodeGeneratorJPG;
 
 class Image extends \Prefab
 {
-    function qrcode($order)
+    function qrcode($order, $rewrite = false)
     {
         $dir = ROOT . '/html/qrcode/';
         if (!is_dir($dir)) {
             mkdir($dir);
         }
-        $image = $dir . $sku . '.png';
-        $code = new QrCode('https://oms.onlymaker.cn/model/ImageView?sku=' . $sku);
-        $code->setWriterByName('png');
-        $code->setMargin(5);
-        $code->setEncoding('UTF-8');
-        $code->setSize(100);
-        $code->setForegroundColor(['r' => 0, 'g' => 0, 'b' => 0, 'a' => 0]);
-        $code->setBackgroundColor(['r' => 255, 'g' => 255, 'b' => 255, 'a' => 0]);
-        $code->setLabel($sku, 10, ROOT . '/html/asset/font/SourceHanSansSC-Light.otf', LabelAlignment::CENTER());
-        //$code->setLogoPath(ROOT . '/html/asset/img/onlymaker.png');
-        //$code->setLogoSize(100, 30);
-        $code->setRoundBlockSize(true);
-        $code->setValidateResult(false);
-        $code->writeFile($image);
+        $domain = \Base::instance()->get('DOMAIN');
+        $image = $dir . $order . '.png';
+        if ($rewrite || !is_file($image)) {
+            $code = new QrCode($domain . '/info/qrcode?code=' . $order);
+            $code->setWriterByName('png');
+            $code->setMargin(5);
+            $code->setEncoding('UTF-8');
+            $code->setSize(150);
+            $code->setForegroundColor(['r' => 0, 'g' => 0, 'b' => 0, 'a' => 0]);
+            $code->setBackgroundColor(['r' => 255, 'g' => 255, 'b' => 255, 'a' => 0]);
+            $code->setLabel($order, 12, ROOT . '/html/font/SourceHanSansSC-Light.otf', LabelAlignment::CENTER());
+            $code->setRoundBlockSize(true);
+            $code->setValidateResult(false);
+            $code->writeFile($image);
+        }
+        return [
+            'path' => $image,
+            'url' => $domain . '/qrcode/' . $order . '.png',
+        ];
     }
 
     function barcode($order)
